+++ message_ix_models/tests/test_workflow.py
@@ -1,21 +1,29 @@
 import re
+from typing import Optional
 
 import ixmp
 import pytest
-from genno import KeyExistsError
 from message_ix import make_df
 
 from message_ix_models import Workflow, testing
-from message_ix_models.workflow import WorkflowStep, solve
+from message_ix_models.workflow import WorkflowStep, make_click_command, solve
+
+MARK = pytest.mark.skipif(
+    condition=ixmp.__version__ < "3.5",
+    reason="ixmp.TimeSeries.url not available prior to ixmp 3.5.0",
+)
 
 
-def changes_a(c, s):
+# Functions for WorkflowSteps
+
+
+def changes_a(c, s) -> None:
     """Change a scenario by modifying structure data, but not data."""
     with s.transact():
         s.add_set("technology", "test_tech")
 
 
-def changes_b(c, s, value=None):
+def changes_b(c, s, value=None) -> None:
     """Change a scenario by modifying parameter data, but not structure."""
     with s.transact():
         s.add_par(
@@ -32,7 +40,7 @@ def changes_b(c, s, value=None):
 
 
 class TestWorkflowStep:
-    def test_call(self, test_context):
+    def test_call(self, test_context) -> None:
         def action(c, s):
             pass  # pragma: no cover
 
@@ -41,25 +49,30 @@ class TestWorkflowStep:
         with pytest.raises(RuntimeError):
             ws(test_context, None)
 
-    def test_repr(self):
+    def test_repr(self) -> None:
         assert "<Step load>" == repr(WorkflowStep(None))
 
 
-@pytest.mark.skipif(
-    condition=ixmp.__version__ < "3.5",
-    reason="ixmp.TimeSeries.url not available prior to ixmp 3.5.0",
-)
-def test_workflow(caplog, request, test_context):
-    # FIXME disentangle this to fewer tests of atomic behaviour
-    base_scenario = testing.bare_res(request, test_context, solved=False)
-    base_url = base_scenario.url
-    base_platform = base_scenario.platform.name
-    del base_scenario
+@pytest.fixture(scope="function")
+def wf(request, test_context) -> Workflow:
+    return _wf(test_context, request=request)
 
-    caplog.clear()
 
+def _wf(
+    context,
+    *,
+    base_url: Optional[str] = None,
+    base_platform: Optional[str] = None,
+    request=None,
+):
+    if base_url is base_platform is None:
+        base_scenario = testing.bare_res(request, context, solved=False)
+        base_platform = base_scenario.platform.name
+        base_url = f"ixmp://{base_platform}/{base_scenario.url}"
+
+    """A function that generates a Workflow."""
     # Create the workflow
-    wf = Workflow(test_context)
+    wf = Workflow(context)
 
     # Model/base is created from nothing by calling base_scenario
     wf.add_step("base", None, target=base_url)
@@ -67,6 +80,53 @@ def test_workflow(caplog, request, test_context):
     wf.add_step("A", "base", changes_a)
     # Model/B is created from Model/A by calling changes_b
     wf.add_step("B", "A", changes_b, value=100.0)
+
+    # Store extra info
+    wf.graph.update({"_base_platform": base_platform})
+
+    return wf
+
+
+@MARK
+def test_make_click_command(mix_models_cli) -> None:
+    import click
+
+    # make_click_command() runs and generates a command
+    name = "make-click-command"
+    cmd = make_click_command(f"{__name__}._wf", name=name, slug="test")
+    assert isinstance(cmd, click.Command)
+
+    # Add this into the hidden CLI test group
+    mix_models_cli.add_command(cmd)
+
+    # Invoke the command with various parameters
+    for params, output in (
+        (["--go", "B"], "nothing returned, workflow will continue with"),
+        (["B"], "Workflow diagram written to"),
+    ):
+        # Command runs and exits with 0
+        result = mix_models_cli.assert_exit_0(["_test", "run"] + params)
+        # Expected log messages or output were printed
+        assert output in result.output
+
+    # Invalid usage
+    for params, output in (
+        (["--go", "C"], "Error: No step(s) matched"),
+        (["--go"], "Error: No target step provided and no default for"),
+        # Step changes_b() fails if changes_a() is not first run
+        (["--go", "--from=[AX]", "B"], "Execute <function changes_b"),
+    ):
+        result = mix_models_cli.invoke(["_test", "run"] + params)
+        assert 0 != result.exit_code
+        assert output in result.output
+
+
+@MARK
+def test_workflow(caplog, request, test_context, wf) -> None:
+    # Retrieve some information from the fixture
+    mp = wf.graph.pop("_base_platform")
+
+    caplog.clear()
 
     # "B solved" is created from "Model/B" by clone and running solve()
     # clone=True without target= raises an exception
@@ -86,7 +146,6 @@ def test_workflow(caplog, request, test_context):
 
     # Log messages reflect workflow steps executed
     start_index = 1 if caplog.messages[0].startswith("Cull") else 0
-    mp = base_platform
     m = "MESSAGEix-GLOBIOM R14 YB"
     messages = [
         f"Loaded ixmp://{mp}/{m}/test_workflow#1",
@@ -120,14 +179,6 @@ def test_workflow(caplog, request, test_context):
     )
 
     # Now truncate the workflow at "Model/A"
-    with pytest.raises(RuntimeError, match="Unable to locate platform info for"):
-        wf.truncate("A")
-
-    # Add a full URL including platform info
-    with pytest.raises(KeyExistsError):
-        wf.add_step("base", None, target=f"ixmp://{base_platform}/{base_url}")
-
-    wf.add_step("base", None, target=f"ixmp://{base_platform}/{base_url}", replace=True)
     wf.truncate("A")
 
     # Description reflects that changes_a() will no longer be called