+++ message_ix_models/tests/tools/test_advance.py
@@ -5,7 +5,6 @@ import pytest
 from message_ix_models.tools.advance import (
     DIMS,
     LOCATION,
-    _fuzz_data,
     advance_data,
     get_advance_data,
 )
@@ -34,16 +33,3 @@ def test_get_advance_data(session_context):
     # Returns a genno.Quantity with the expected units
     result = advance_data("Transport|Service demand|Road|Freight")
     assert {"[length]": 1, "[mass]": 1, "[time]": -1} == result.units.dimensionality
-
-
-def test_fuzz_data(session_context):
-    """Test that :func:`_fuzz_data` runs successfully.
-
-    NB this only produces a file in the :mod:`pytest` temporary directory. To update
-    the test specimen in the package directory that is used by
-    :func:`test_get_advance_data`, see the body of _fuzz_data.
-    """
-    # size argument should be a fraction (~= 0.1) of the size of the test specimen
-    _fuzz_data(
-        size=50, include=[("Transport|Service demand|Road|Freight", "billion tkm/yr")]
-    )